move inspect approvers to supervision/inspect_ai
