--- el/src/el/approvers/_registry.py
@@ -1,1 +1,0 @@
-from .approvers import bash_approver, python_approver, human_approver, llm_approver
