+++ el/src/el/supervision/config.py
@@ -1,0 +1,72 @@
+from typing import Any, Callable, Dict, List, Optional, Literal
+from enum import Enum
+import random
+import json
+from pydantic import BaseModel, Field
+
+class SupervisionDecisionType(str, Enum):
+    APPROVE = "approve"
+    REJECT = "reject"
+    ESCALATE = "escalate"
+    TERMINATE = "terminate"
+    MODIFY = "modify"
+
+class SupervisionDecision(BaseModel):
+    decision: SupervisionDecisionType
+    """Supervision decision."""
+
+    modified: Any = Field(default=None)
+    """Modified data for decision 'modify'."""
+
+    explanation: str | None = Field(default=None)
+    """Explanation for decision."""
+
+class SupervisionConfig:
+    def __init__(self):
+        self.global_supervision_functions: List[Callable] = []
+        self.global_mock_policy = None
+        self.override_local_policy = False
+        self.mock_responses: Dict[str, List[Any]] = {}
+        self.previous_calls: Dict[str, List[Any]] = {}
+        self.function_supervisors: Dict[str, List[Callable]] = {}  # Function-specific supervision chains
+
+    def set_global_supervision_functions(self, functions: List[Callable]):
+        self.global_supervision_functions = functions
+
+    def set_function_supervision_functions(self, function_name: str, functions: List[Callable]):
+        self.function_supervisors[function_name] = functions
+
+    def get_supervision_functions(self, function_name: str, local_supervision: Optional[List[Callable]] = None) -> List[Callable]:
+        supervisors = []
+        # Add local supervision functions specified in the decorator
+        if local_supervision:
+            supervisors.extend(local_supervision)
+        
+        # Add function-specific supervision functions
+        local_supervisors = self.function_supervisors.get(function_name, [])
+        for supervisor in local_supervisors:
+            if supervisor not in supervisors:
+                supervisors.append(supervisor)
+        
+        # Add global supervision functions
+        for supervisor in self.global_supervision_functions:
+            if supervisor not in supervisors:
+                supervisors.append(supervisor)
+        
+        return supervisors
+
+    def add_function_supervisors(self, function_name: str, supervisors: List[Callable]):
+        """Add supervisor functions to the supervision chain of a specific function."""
+        if function_name not in self.function_supervisors:
+            self.function_supervisors[function_name] = []
+        self.function_supervisors[function_name].extend(supervisors)
+
+# Global instance of SupervisionConfig
+# TODO: Update this
+supervision_config = SupervisionConfig()
+
+def set_global_supervision_functions(functions: List[Callable]):
+    supervision_config.set_global_supervision_functions(functions)
+
+def set_function_supervision_functions(function_name: str, functions: List[Callable]):
+    supervision_config.set_function_supervision_functions(function_name, functions)
