+++ el/src/el/supervision/inspect_ai/__init__.py
@@ -1,0 +1,3 @@
+from .supervisors import bash_approver, python_approver, human_approver, llm_approver
+
+__all__ = ["bash_approver", "python_approver", "human_approver", "llm_approver"]
