+++ el/src/el/supervision/inspect_ai/_config.py
@@ -1,0 +1,4 @@
+# Constants
+DEFAULT_TIMEOUT = 300
+DEFAULT_SUGGESTIONS = 5
+SLEEP_TIME = 2
