+++ el/src/el/supervision/inspect_ai/_registry.py
@@ -1,0 +1,1 @@
+from .supervisors import bash_approver, python_approver, human_approver, llm_approver
